# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Waypoint Delivery Planning System API
  description: OpenAPI 3.1 specification for the Waypoint Group delivery planning system.
  contact:
    email: tech-triathlon@rootcode.io
  version: 1.0.0
servers:
  - url: /api/v1
    description: Production API Base Endpoint
...
paths:
...
  /loading/trips:
    get:
      operationId: getLoadingTrips
      summary: Get confirmed trips ready for warehouse loading
      tags:
        - loader
      x-role:
        - loader
      security:
        - BearerAuth: []
      parameters:
        - name: depot_id
          in: query
          required: false
          schema:
            type: string
          example: Peliyagoda
      responses:
        '200':
          description: Confirmed trip cards for loading
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TripCard'
              example:
                - id: TRIP-20250801-01
                  plan_run_id: RUN-20250801-PEL
                  delivery_date: "2025-08-01"
                  trip_no: 1
                  brand: Fresh
                  district: Colombo
                  depot_id: Peliyagoda
                  vehicle_id: VEH001
                  vehicle_type: truck
                  vehicle_temp: chilled
                  stop_count: 1
                  depart_time: "03:30"
                  plan_version: 1
                  status: CONFIRMED
                  weight_used_kg: 180.0
                  weight_cap_kg: 3610.0
                  volume_used_m3: 0.75
                  volume_cap_m3: 19.4
  /trips/{id}/load-list:
    get:
      operationId: getTripLoadList
      summary: Get trip load manifest with delivery sequence and reverse load order
      tags:
        - loader
      x-role:
        - loader
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: TRIP-20250801-01
      responses:
        '200':
          description: Load manifest details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoadListResponse'
              example:
                plan_version: 1
                trip:
                  id: TRIP-20250801-01
                  plan_run_id: RUN-20250801-PEL
                  delivery_date: "2025-08-01"
                  trip_no: 1
                  brand: Fresh
                  district: Colombo
                  depot_id: Peliyagoda
                  vehicle_id: VEH001
                  vehicle_type: truck
                  vehicle_temp: chilled
                  stop_count: 1
                  depart_time: "03:30"
                  plan_version: 1
                  status: CONFIRMED
                  weight_used_kg: 180.0
                  weight_cap_kg: 3610.0
                  volume_used_m3: 0.75
                  volume_cap_m3: 19.4
                delivery_sequence: []
                reverse_load_order: []
  /trips/{id}/load-start:
    post:
      operationId: startTripLoading
      summary: Start warehouse loading for a confirmed trip
      tags:
        - loader
      x-role:
        - loader
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: TRIP-20250801-01
      responses:
        '200':
          description: Trip transitioning to LOADING status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trip'
        '409':
          description: Plan changed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /trips/{id}/load-checks:
    post:
      operationId: reportLoadCheck
      summary: Report missing or damaged stock loading shortfall
      tags:
        - loader
      x-role:
        - loader
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: TRIP-20250801-01
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLoadCheckRequest'
            example:
              order_id: ORD-20250801-001
              plan_version: 1
              expected_qty: 150
              loaded_qty: 140
              issue: missing
              note: 10 units missing from staging bay.
              client_op_id: e102911a-0012-4211-b012-aa7719281100
      responses:
        '201':
          description: Reported shortfall
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoadCheck'
        '409':
          description: Stale plan version
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /trips/{id}/load-confirm:
    post:
      operationId: confirmTripLoad
      summary: Confirm trip loading completed at current plan_version
      tags:
        - loader
      x-role:
        - loader
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: TRIP-20250801-01
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmLoadRequest'
            example:
              plan_version: 1
      responses:
        '200':
          description: Trip loaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trip'
        '409':
          description: Open shortfall or stale plan version blocking confirm
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ----------------------------------------------------
  # Driver Endpoints
  # ----------------------------------------------------
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer authentication header. Role and scope are derived from the token.
  schemas:
...
    Brand:
      type: string
      enum:
        - Fresh
        - Style
        - Tech
      description: Operating retail brand of Waypoint Group.
...
    TemperatureRequirement:
      type: string
      enum:
        - chilled
        - ambient
      description: Temperature requirement for orders or vehicles.
    VehicleType:
      type: string
      enum:
        - truck
        - van
      description: Classification of vehicle.
    DockType:
      type: string
      enum:
        - rear_dock
        - street
        - mall_bay
      description: Unloading facility type at an outlet.
    ParkingConstraint:
      type: string
      enum:
        - normal
        - van_only
        - mall_dock
      description: Access constraint for parking/unloading at an outlet.
...
    TripStatus:
      type: string
      enum:
        - DRAFT
        - CONFIRMED
        - LOADING
        - LOADED
        - IN_PROGRESS
        - COMPLETED
        - BLOCKED
        - CANCELLED
      description: Status of a delivery trip.
    StopStatus:
      type: string
      enum:
        - PENDING
        - ARRIVED
        - DELIVERED
        - PARTIAL
        - FAILED
        - EXCEPTION
        - SKIPPED
      description: Status of an individual trip stop.
    ReceiptStatus:
      type: string
      enum:
        - NONE
        - AWAITING
        - CONFIRMED
        - DISCREPANCY
      description: Store manager receipt confirmation status for a stop.
    StopOutcome:
      type: string
      enum:
        - delivered
        - partial
        - refused
        - closed
      description: Outcome recorded by the driver at a stop.
...
    LoadCheckIssue:
      type: string
      enum:
        - missing
        - damaged
      description: Issue reported during loading check.
    LoadCheckStatus:
      type: string
      enum:
        - OPEN
        - RESOLVED
      description: Status of a load check shortfall report.
    LoadCheckResolution:
      type: string
      enum:
        - defer_order
        - replan_order
        - proceed_partial
      description: Dispatcher resolution for a loading shortfall.
...
    Violation:
      type: object
      required:
        - rule
        - message
      properties:
        rule:
          type: string
          description: Rule ID (e.g. H1 through H12).
        message:
          type: string
          description: Human-readable description of rule violation.
        actual:
          description: Observed value that violated constraint.
        limit:
          description: Threshold limit specified by the rule.
      description: Constraint violation detail item.
    ErrorDetails:
      type: object
      properties:
        violations:
          type: array
          items:
            $ref: '#/components/schemas/Violation'
          description: List of rule violations if code is CONSTRAINT_VIOLATION.
      additionalProperties: true
      description: Detailed contextual metadata for error payload.
    ErrorBody:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Error code (e.g. CONSTRAINT_VIOLATION, PLAN_CHANGED, NOT_OPERATING_DAY, INVALID_TRANSITION, FORBIDDEN_SCOPE, OPEN_SHORTFALL, STALE_PLAN_VERSION).
        message:
          type: string
          description: High-level error message.
        details:
          $ref: '#/components/schemas/ErrorDetails'
      description: Core error payload.
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
      description: Standard API error response envelope.

    # ----------------------------------------------------
    # Domain Model & DTO Schemas
    # ----------------------------------------------------
...
    Trip:
      type: object
      required:
        - id
        - plan_run_id
        - vehicle_id
        - depot_id
        - delivery_date
        - trip_no
        - brand
        - district
        - status
        - plan_version
        - est_minutes
        - est_km
        - est_fuel_l
      properties:
        id:
          type: string
          description: Trip identifier.
        plan_run_id:
          type: string
          description: Associated planning run ID.
        vehicle_id:
          type: string
          description: Assigned vehicle code.
        depot_id:
          type: string
          description: Depot code.
        delivery_date:
          type: string
          format: date
          description: Date of execution.
        trip_no:
          type: integer
          enum: [1, 2]
          description: Trip sequence number for vehicle on date (1 or 2).
        brand:
          $ref: '#/components/schemas/Brand'
        district:
          type: string
          description: Target district served by trip.
        status:
          $ref: '#/components/schemas/TripStatus'
        plan_version:
          type: integer
          description: Current revision version of trip plan.
        depart_time:
          type: string
          nullable: true
          description: Planned departure time (HH:MM).
        est_minutes:
          type: integer
          description: Planned trip duration in minutes.
        est_km:
          type: number
          description: Total estimated travel distance in km.
        est_fuel_l:
          type: number
          description: Estimated fuel consumption in litres.
        confirmed_by:
          type: string
          nullable: true
          description: Dispatcher ID who confirmed the trip.
        confirmed_at:
          type: string
          format: date-time
          nullable: true
          description: Confirmation timestamp.
      description: Delivery trip record.
    TripCard:
      type: object
      required:
        - id
        - plan_run_id
        - delivery_date
        - trip_no
        - brand
        - district
        - depot_id
        - vehicle_id
        - vehicle_type
        - vehicle_temp
        - stop_count
        - plan_version
        - status
        - weight_used_kg
        - weight_cap_kg
        - volume_used_m3
        - volume_cap_m3
      properties:
        id:
          type: string
        plan_run_id:
          type: string
        delivery_date:
          type: string
          format: date
        trip_no:
          type: integer
          enum: [1, 2]
        brand:
          $ref: '#/components/schemas/Brand'
        district:
          type: string
        depot_id:
          type: string
        vehicle_id:
          type: string
        vehicle_type:
          $ref: '#/components/schemas/VehicleType'
        vehicle_temp:
          $ref: '#/components/schemas/TemperatureRequirement'
        stop_count:
          type: integer
        depart_time:
          type: string
          nullable: true
        plan_version:
          type: integer
        status:
          $ref: '#/components/schemas/TripStatus'
        weight_used_kg:
          type: number
        weight_cap_kg:
          type: number
        volume_used_m3:
          type: number
        volume_cap_m3:
          type: number
        est_minutes:
          type: integer
          nullable: true
          example: 112
        est_fuel_l:
          type: number
          nullable: true
          example: 12.4
        time_budget_min:
          type: integer
          enum: [270, 480]
          nullable: true
          example: 270
      description: Lightweight trip summary object used in loader and driver lists (D23).
    StopDetail:
      type: object
      required:
        - id
        - trip_id
        - order_id
        - seq
        - outlet_id
        - outlet_name
        - district
        - dock_type
        - parking_constraint
        - order_units
        - order_weight_kg
        - order_volume_m3
        - window_open_time
        - window_close_time
        - service_min
        - status
        - receipt_status
      properties:
        id:
          type: string
        trip_id:
          type: string
        order_id:
          type: string
        seq:
          type: integer
        outlet_id:
          type: string
        outlet_name:
          type: string
          example: Fresh Colombo 03
        district:
          type: string
        dock_type:
          $ref: '#/components/schemas/DockType'
        parking_constraint:
          $ref: '#/components/schemas/ParkingConstraint'
        order_units:
          type: integer
        order_weight_kg:
          type: number
        order_volume_m3:
          type: number
        window_open_time:
          type: string
        window_close_time:
          type: string
        eta:
          type: string
          nullable: true
        service_start_est:
          type: string
          nullable: true
          example: "04:15"
        service_min:
          type: integer
        status:
          $ref: '#/components/schemas/StopStatus'
        receipt_status:
          $ref: '#/components/schemas/ReceiptStatus'
        arrived_at:
          type: string
          format: date-time
          nullable: true
        completed_at:
          type: string
          format: date-time
          nullable: true
        outcome:
          $ref: '#/components/schemas/StopOutcome'
          nullable: true
        quantity_delivered:
          type: integer
          nullable: true
        received_by:
          type: string
          nullable: true
        outcome_note:
          type: string
          nullable: true
      description: Detailed stop payload with embedded outlet and order attributes (D23).
...
    LoadCheck:
      type: object
      required:
        - id
        - trip_id
        - order_id
        - plan_version
        - expected_qty
        - loaded_qty
        - issue
        - status
        - reported_by
      properties:
        id:
          type: string
          description: Load check shortfall record ID.
        trip_id:
          type: string
          description: Trip ID being loaded.
        order_id:
          type: string
          description: Affected order ID.
        plan_version:
          type: integer
          description: Plan version active when reported.
        expected_qty:
          type: integer
          description: Expected unit count from manifest.
        loaded_qty:
          type: integer
          description: Actual physical count loaded.
        issue:
          $ref: '#/components/schemas/LoadCheckIssue'
        note:
          type: string
          nullable: true
          description: Loader explanation note.
        status:
          $ref: '#/components/schemas/LoadCheckStatus'
        resolution:
          $ref: '#/components/schemas/LoadCheckResolution'
          nullable: true
        reported_by:
          type: string
          description: Loader user ID.
        resolved_by:
          type: string
          nullable: true
          description: Dispatcher user ID resolving shortfall.
        vehicle_id:
          type: string
          nullable: true
          example: VEH001
        outlet_id:
          type: string
          nullable: true
          example: OUT004
        outlet_name:
          type: string
          nullable: true
          example: Fresh Store Colombo 04
        bay:
          type: string
          nullable: true
          example: Bay 2
      description: Warehouse loading shortfall audit record.
...
    CreateLoadCheckRequest:
      type: object
      required:
        - order_id
        - plan_version
        - expected_qty
        - loaded_qty
        - issue
      properties:
        order_id:
          type: string
        plan_version:
          type: integer
        expected_qty:
          type: integer
        loaded_qty:
          type: integer
        issue:
          $ref: '#/components/schemas/LoadCheckIssue'
        note:
          type: string
        client_op_id:
          type: string
          format: uuid
      description: Loader shortfall reporting payload.
    ConfirmLoadRequest:
      type: object
      required:
        - plan_version
      properties:
        plan_version:
          type: integer
          description: Must match current trip plan version.
      description: Loader confirmation payload.
...
    LoadListResponse:
      type: object
      required:
        - plan_version
        - trip
        - delivery_sequence
        - reverse_load_order
      properties:
        plan_version:
          type: integer
        trip:
          $ref: '#/components/schemas/TripCard'
        delivery_sequence:
          type: array
          items:
            $ref: '#/components/schemas/StopDetail'
        reverse_load_order:
          type: array
          items:
            $ref: '#/components/schemas/StopDetail'
      description: Warehouse load list response containing manifest sequence and reverse load order (D23).